inverted mask
- 1 in the mask now means "opaque"

This is so that shifting the mask before applying it doesn't change the
mask.  Then 0 was opaque shifting the mask would add more 0s and could
produce expected results.  Of course you could one compliment it twice
before, but then you lose more speed.  So we've gone back to 1 meaning
opaque for now - and only a single bit flip is needed to use the mask
with AND.